iox-#1293 Move impl's back to public dir to show their doxygen in the docs
