+++ iceoryx_binding_c/include/iceoryx_binding_c/internal/c2cpp_enum_translation.hpp
@@ -19,7 +19,7 @@
 
 #include "c2cpp_binding.h"
 #include "iceoryx_binding_c/enums.h"
-#include "iceoryx_posh/internal/popo/base_subscriber.hpp"
+#include "iceoryx_posh/popo/impl/base_subscriber.hpp"
 #include "iceoryx_posh/popo/subscriber.hpp"
 #include "iceoryx_posh/runtime/service_discovery.hpp"
