+++ iceoryx_posh/include/iceoryx_posh/internal/popo/subscriber_impl.inl
@@ -18,7 +18,7 @@
 #ifndef IOX_POSH_POPO_TYPED_SUBSCRIBER_IMPL_INL
 #define IOX_POSH_POPO_TYPED_SUBSCRIBER_IMPL_INL
 
-#include "iceoryx_posh/internal/popo/subscriber_impl.hpp"
+#include "iceoryx_posh/popo/impl/subscriber_impl.hpp"
 
 namespace iox
 {
