+++ iceoryx_posh/include/iceoryx_posh/internal/popo/untyped_client_impl.inl
@@ -17,7 +17,7 @@
 #ifndef IOX_POSH_POPO_UNTYPED_CLIENT_IMPL_INL
 #define IOX_POSH_POPO_UNTYPED_CLIENT_IMPL_INL
 
-#include "iceoryx_posh/internal/popo/untyped_client_impl.hpp"
+#include "iceoryx_posh/popo/impl/untyped_client_impl.hpp"
 
 namespace iox
 {
