+++ iceoryx_posh/include/iceoryx_posh/popo/impl/publisher_impl.hpp
@@ -19,9 +19,9 @@
 #define IOX_POSH_POPO_TYPED_PUBLISHER_IMPL_HPP
 
 #include "iceoryx_hoofs/cxx/type_traits.hpp"
-#include "iceoryx_posh/internal/popo/base_publisher.hpp"
 #include "iceoryx_posh/internal/popo/publisher_interface.hpp"
 #include "iceoryx_posh/internal/popo/typed_port_api_trait.hpp"
+#include "iceoryx_posh/popo/impl/base_publisher.hpp"
 #include "iceoryx_posh/popo/sample.hpp"
 
 namespace iox
