+++ iceoryx_posh/include/iceoryx_posh/popo/impl/server_impl.hpp
@@ -18,11 +18,11 @@
 #define IOX_POSH_POPO_SERVER_IMPL_HPP
 
 #include "iceoryx_posh/capro/service_description.hpp"
-#include "iceoryx_posh/internal/popo/base_server.hpp"
 #include "iceoryx_posh/internal/popo/request_deleter.hpp"
 #include "iceoryx_posh/internal/popo/response_deleter.hpp"
 #include "iceoryx_posh/internal/popo/rpc_interface.hpp"
 #include "iceoryx_posh/internal/popo/typed_port_api_trait.hpp"
+#include "iceoryx_posh/popo/impl/base_server.hpp"
 #include "iceoryx_posh/popo/request.hpp"
 #include "iceoryx_posh/popo/response.hpp"
 #include "iceoryx_posh/popo/server_options.hpp"
