+++ iceoryx_posh/include/iceoryx_posh/popo/impl/untyped_subscriber_impl.hpp
@@ -22,7 +22,7 @@
 #include "iceoryx_hoofs/cxx/unique_ptr.hpp"
 #include "iceoryx_posh/capro/service_description.hpp"
 #include "iceoryx_posh/iceoryx_posh_types.hpp"
-#include "iceoryx_posh/internal/popo/base_subscriber.hpp"
+#include "iceoryx_posh/popo/impl/base_subscriber.hpp"
 
 namespace iox
 {
