+++ iceoryx_posh/include/iceoryx_posh/popo/untyped_client.hpp
@@ -17,7 +17,7 @@
 #ifndef IOX_POSH_POPO_UNTYPED_CLIENT_HPP
 #define IOX_POSH_POPO_UNTYPED_CLIENT_HPP
 
-#include "iceoryx_posh/internal/popo/untyped_client_impl.hpp"
+#include "iceoryx_posh/popo/impl/untyped_client_impl.hpp"
 
 namespace iox
 {
