+++ iceoryx_posh/include/iceoryx_posh/popo/untyped_server.hpp
@@ -17,7 +17,7 @@
 #ifndef IOX_POSH_POPO_UNTYPED_SERVER_HPP
 #define IOX_POSH_POPO_UNTYPED_SERVER_HPP
 
-#include "iceoryx_posh/internal/popo/untyped_server_impl.hpp"
+#include "iceoryx_posh/popo/impl/untyped_server_impl.hpp"
 
 namespace iox
 {
