+++ iceoryx_posh/test/mocks/publisher_mock.hpp
@@ -19,7 +19,7 @@
 #define IOX_POSH_MOCKS_PUBLISHER_MOCK_HPP
 
 #include "iceoryx_hoofs/cxx/expected.hpp"
-#include "iceoryx_posh/internal/popo/base_publisher.hpp"
+#include "iceoryx_posh/popo/impl/base_publisher.hpp"
 #include "iceoryx_posh/popo/publisher.hpp"
 #include "iceoryx_posh/popo/sample.hpp"
