+++ iceoryx_posh/test/moduletests/test_popo_base_server.cpp
@@ -15,9 +15,9 @@
 // SPDX-License-Identifier: Apache-2.0
 
 #include "iceoryx_hoofs/cxx/optional.hpp"
-#include "iceoryx_posh/internal/popo/base_server.hpp"
-#include "iceoryx_posh/internal/popo/server_impl.hpp"
-#include "iceoryx_posh/internal/popo/untyped_server_impl.hpp"
+#include "iceoryx_posh/popo/impl/base_server.hpp"
+#include "iceoryx_posh/popo/impl/server_impl.hpp"
+#include "iceoryx_posh/popo/impl/untyped_server_impl.hpp"
 #include "iceoryx_posh/testing/mocks/posh_runtime_mock.hpp"
 #include "mocks/server_mock.hpp"
 #include "mocks/trigger_handle_mock.hpp"
