+++ iceoryx_posh/test/moduletests/test_popo_base_subscriber.cpp
@@ -19,8 +19,8 @@
 #include "iceoryx_hoofs/cxx/optional.hpp"
 #include "iceoryx_hoofs/cxx/unique_ptr.hpp"
 #include "iceoryx_posh/iceoryx_posh_types.hpp"
-#include "iceoryx_posh/internal/popo/base_subscriber.hpp"
 #include "iceoryx_posh/internal/popo/building_blocks/condition_variable_data.hpp"
+#include "iceoryx_posh/popo/impl/base_subscriber.hpp"
 #include "iceoryx_posh/popo/wait_set.hpp"
 
 #include "iceoryx_posh/testing/mocks/chunk_mock.hpp"
